requirements.txt
tensorboard>=2.15.0
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
tqdm>=4.50.0
onnxruntime-gpu>=1.19.2
nemo-toolkit[asr]
transformers>=4.21.0
librosa>=0.9.0
soundfile>=0.10.0